+++ packages/sdk-redux/src/reduxSlices/rtkQuery/rpcApiSlice/endpoints/flowEndpoints.ts
@@ -26,7 +26,7 @@ export const createFlowEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: senderAddress,
+                signerAddress: senderAddress,
                 dispatch: queryApi.dispatch,
                 title: 'Create Stream',
                 extraData: arg.transactionExtraData,
@@ -52,7 +52,7 @@ export const createFlowEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: senderAddress,
+                signerAddress: senderAddress,
                 dispatch: queryApi.dispatch,
                 title: 'Update Stream',
                 extraData: arg.transactionExtraData,
@@ -78,7 +78,7 @@ export const createFlowEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: senderAddress,
+                signerAddress: senderAddress,
                 dispatch: queryApi.dispatch,
                 title: 'Close Stream',
                 extraData: arg.transactionExtraData,