+++ packages/sdk-redux/src/reduxSlices/rtkQuery/rpcApiSlice/endpoints/indexEndpoints.ts
@@ -30,7 +30,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Create Index',
                 extraData: arg.transactionExtraData,
@@ -54,7 +54,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Distribute Index',
                 extraData: arg.transactionExtraData,
@@ -79,7 +79,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Update Index Subscription Units',
                 extraData: arg.transactionExtraData,
@@ -103,7 +103,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Approve Index Subscription',
                 extraData: arg.transactionExtraData,
@@ -128,7 +128,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Claim from Index Subscription',
                 extraData: arg.transactionExtraData,
@@ -153,7 +153,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Delete Index Subscription',
                 extraData: arg.transactionExtraData,
@@ -177,7 +177,7 @@ export const createIndexEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Revoke Index Subscription',
                 extraData: arg.transactionExtraData,