+++ packages/sdk-redux/src/reduxSlices/rtkQuery/rpcApiSlice/endpoints/superTokenEndpoints.ts
@@ -35,7 +35,7 @@ export const createSuperTokenEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Upgrade to Super Token',
                 extraData: arg.transactionExtraData,
@@ -86,7 +86,7 @@ export const createSuperTokenEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Downgrade from Super Token',
                 extraData: arg.transactionExtraData,
@@ -109,7 +109,7 @@ export const createSuperTokenEndpoints = (builder: RpcEndpointBuilder) => ({
             return await registerNewTransactionAndReturnQueryFnResult({
                 transactionResponse,
                 chainId: arg.chainId,
-                signer: await arg.signer.getAddress(),
+                signerAddress: await arg.signer.getAddress(),
                 dispatch: queryApi.dispatch,
                 title: 'Transfer Super Token',
                 extraData: arg.transactionExtraData,