+++ packages/sdk-redux/src/reduxSlices/transactionTrackerSlice/trackedTransaction.ts
@@ -11,7 +11,7 @@ export interface TrackedTransaction {
     /**
      * The address this transaction is from.
      */
-    signer: string;
+    signerAddress: string;
     /**
      * Milliseconds since epoch when started tracking the transaction.
      */